# openapi.yaml
openapi: 3.0.3
info:
  title: B2B Stock & Orders API
  version: "1.0"
servers:
  - url: https://k7n8794nyf.execute-api.eu-north-1.amazonaws.com/v1
paths:
  /orders:
    post:
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [companyId, quantity]
              properties:
                companyId: { type: string, example: "COMP1" }
                quantity: { type: number, example: 5 }
      responses:
        "201": { description: Created }
        "400": { description: Bad Request }
  /orders/company/{companyId}:
    get:
      summary: List company orders
      parameters:
        - name: companyId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
...